Remove ethereumAdapters and Safe contracts from safe-core-sdk
