+++ packages/safe-core-sdk/package.json
@@ -11,9 +11,8 @@
     "SDK"
   ],
   "scripts": {
-    "typechain": "ts-node scripts/generateTypechainFiles.ts",
-    "unbuild": "rimraf dist artifacts cache .nyc_output typechain",
-    "build": "hardhat compile && yarn typechain && tsc",
+    "unbuild": "rimraf dist artifacts cache .nyc_output",
+    "build": "hardhat compile && tsc",
     "test:ganache:web3:v1.2.0": "export TEST_NETWORK=ganache && export ETH_LIB=web3 && export SAFE_VERSION=1.2.0 && hardhat --network localhost deploy && nyc hardhat --network localhost test",
     "test:ganache:web3:v1.3.0": "export TEST_NETWORK=ganache && export ETH_LIB=web3 && export SAFE_VERSION=1.3.0 && hardhat --network localhost deploy && nyc hardhat --network localhost test",
     "test:ganache:ethers:v1.2.0": "export TEST_NETWORK=ganache && export ETH_LIB=ethers && export SAFE_VERSION=1.2.0 && hardhat --network localhost deploy && nyc hardhat --network localhost test",
@@ -40,43 +39,41 @@
   ],
   "homepage": "https://github.com/gnosis/safe-core-sdk#readme",
   "devDependencies": {
+    "@gnosis.pm/safe-ethers-lib": "^1.0.0",
+    "@gnosis.pm/safe-web3-lib": "^1.0.0",
     "@gnosis.pm/safe-contracts-v1.2.0": "npm:@gnosis.pm/safe-contracts@1.2.0",
     "@gnosis.pm/safe-contracts-v1.3.0": "npm:@gnosis.pm/safe-contracts@1.3.0",
-    "@nomiclabs/hardhat-ethers": "^2.0.3",
-    "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "@nomiclabs/hardhat-ethers": "^2.0.4",
+    "@nomiclabs/hardhat-waffle": "^2.0.2",
     "@nomiclabs/hardhat-web3": "^2.0.0",
-    "@typechain/ethers-v5": "^8.0.5",
-    "@typechain/web3-v1": "^3.0.0",
     "@types/chai": "^4.3.0",
     "@types/chai-as-promised": "^7.1.4",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^17.0.8",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.10",
     "@types/semver": "^7.3.9",
     "@types/yargs": "^16.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.9.0",
-    "@typescript-eslint/parser": "^5.9.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
     "chai": "^4.3.4",
     "chai-as-promised": "^7.1.1",
     "coveralls": "^3.1.1",
     "dotenv": "^10.0.0",
-    "eslint": "^8.6.0",
+    "eslint": "^8.7.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "ethereum-waffle": "^3.4.0",
-    "ethers": "^5.5.2",
-    "hardhat": "^2.8.0",
+    "ethers": "^5.5.3",
+    "hardhat": "^2.8.3",
     "hardhat-deploy": "^0.9.24",
     "husky": "^7.0.4",
-    "lint-staged": "^12.1.5",
-    "mocha": "^9.1.3",
+    "lint-staged": "^12.2.1",
+    "mocha": "^9.1.4",
     "nyc": "^15.1.0",
     "prettier": "^2.5.1",
     "ts-generator": "^0.1.1",
-    "ts-node": "^10.4.0",
-    "typechain": "^6.0.5",
     "typescript": "^4.5.4",
-    "web3": "^1.6.1",
-    "yargs": "^17.3.0"
+    "web3": "^1.7.0",
+    "yargs": "^17.3.1"
   },
   "lint-staged": {
     "src/**/!(*test).ts": [
@@ -90,6 +87,7 @@
     }
   },
   "dependencies": {
+    "@ethersproject/solidity": "^5.5.0",
     "@gnosis.pm/safe-core-sdk-types": "^0.1.1",
     "@gnosis.pm/safe-deployments": "^1.7.0",
     "ethereumjs-util": "^7.1.3",
