--- packages/safe-core-sdk/src/contracts/GnosisSafe/v1.1.1/GnosisSafeContract_V1_1_1_Ethers.ts
@@ -1,23 +1,0 @@
-import { GnosisSafe } from '../../../../typechain/src/ethers-v5/v1.1.1/GnosisSafe'
-import { sameString } from '../../../utils'
-import GnosisSafeContractEthers from '../GnosisSafeContractEthers'
-
-class GnosisSafeContract_V1_1_1_Ethers extends GnosisSafeContractEthers {
-  constructor(public contract: GnosisSafe) {
-    super(contract)
-  }
-
-  async getModules(): Promise<string[]> {
-    return this.contract.getModules()
-  }
-
-  async isModuleEnabled(moduleAddress: string): Promise<boolean> {
-    const modules = await this.getModules()
-    const isModuleEnabled = modules.some((enabledModuleAddress: string) =>
-      sameString(enabledModuleAddress, moduleAddress)
-    )
-    return isModuleEnabled
-  }
-}
-
-export default GnosisSafeContract_V1_1_1_Ethers
