--- packages/safe-core-sdk/src/contracts/GnosisSafe/v1.2.0/GnosisSafeContract_V1_2_0_Ethers.ts
@@ -1,18 +1,0 @@
-import { GnosisSafe } from '../../../../typechain/src/ethers-v5/v1.2.0/GnosisSafe'
-import GnosisSafeContractEthers from '../GnosisSafeContractEthers'
-
-class GnosisSafeContract_V1_2_0_Ethers extends GnosisSafeContractEthers {
-  constructor(public contract: GnosisSafe) {
-    super(contract)
-  }
-
-  async getModules(): Promise<string[]> {
-    return this.contract.getModules()
-  }
-
-  async isModuleEnabled(moduleAddress: string): Promise<boolean> {
-    return this.contract.isModuleEnabled(moduleAddress)
-  }
-}
-
-export default GnosisSafeContract_V1_2_0_Ethers
