--- packages/safe-core-sdk/src/contracts/GnosisSafeProxyFactory/v1.1.1/GnosisSafeProxyFactoryContract_V1_1_1_Ethers.ts
@@ -1,10 +1,0 @@
-import { ProxyFactory } from '../../../../typechain/src/ethers-v5/v1.1.1/ProxyFactory'
-import GnosisSafeProxyFactoryEthersContract from '../GnosisSafeProxyFactoryEthersContract'
-
-class GnosisSafeProxyFactoryContract_V1_1_1_Ethers extends GnosisSafeProxyFactoryEthersContract {
-  constructor(public contract: ProxyFactory) {
-    super(contract)
-  }
-}
-
-export default GnosisSafeProxyFactoryContract_V1_1_1_Ethers
