--- packages/safe-core-sdk/src/contracts/GnosisSafeProxyFactory/v1.3.0/GnosisSafeProxyFactoryContract_V1_3_0_Web3.ts
@@ -1,10 +1,0 @@
-import { ProxyFactory } from '../../../../typechain/src/web3-v1/v1.3.0/proxy_factory'
-import GnosisSafeProxyFactoryWeb3Contract from '../GnosisSafeProxyFactoryWeb3Contract'
-
-class GnosisSafeProxyFactoryContract_V1_3_0_Web3 extends GnosisSafeProxyFactoryWeb3Contract {
-  constructor(public contract: ProxyFactory) {
-    super(contract)
-  }
-}
-
-export default GnosisSafeProxyFactoryContract_V1_3_0_Web3
