--- packages/safe-core-sdk/src/contracts/MultiSend/MultiSendContract.ts
@@ -1,6 +1,0 @@
-interface MultiSendContract {
-  getAddress(): string
-  encode(methodName: any, params: any): string
-}
-
-export default MultiSendContract
