+++ packages/safe-core-sdk/src/index.ts
@@ -1,7 +1,3 @@
-import { SafeVersion } from './contracts/config'
-import EthAdapter, { EthAdapterTransaction } from './ethereumLibs/EthAdapter'
-import EthersAdapter, { EthersAdapterConfig } from './ethereumLibs/EthersAdapter'
-import Web3Adapter, { Web3AdapterConfig } from './ethereumLibs/Web3Adapter'
 import ContractManager from './managers/contractManager'
 import Safe, {
   AddOwnerTxParams,
@@ -17,32 +13,19 @@ import SafeFactory, {
 } from './safeFactory'
 import { ContractNetworksConfig } from './types'
 import EthSignSignature from './utils/signatures/SafeSignature'
-import {
-  SafeTransactionOptionalProps,
-  TransactionOptions,
-  TransactionResult
-} from './utils/transactions/types'
+import { SafeTransactionOptionalProps } from './utils/transactions/types'
 import { standardizeSafeTransactionData } from './utils/transactions/utils'
 
 export default Safe
 export {
   ContractManager,
-  SafeVersion,
   SafeFactory,
   SafeFactoryConfig,
   SafeAccountConfig,
   SafeDeploymentConfig,
-  EthAdapter,
-  EthAdapterTransaction,
-  Web3AdapterConfig,
-  Web3Adapter,
-  EthersAdapterConfig,
-  EthersAdapter,
   SafeConfig,
   ConnectSafeConfig,
   ContractNetworksConfig,
-  TransactionOptions,
-  TransactionResult,
   SafeTransactionOptionalProps,
   AddOwnerTxParams,
   RemoveOwnerTxParams,
