+++ packages/safe-core-sdk/src/managers/contractManager.ts
@@ -1,6 +1,9 @@
-import { SafeVersion, SAFE_LAST_VERSION } from '../contracts/config'
-import GnosisSafeContract from '../contracts/GnosisSafe/GnosisSafeContract'
-import MultiSendContract from '../contracts/MultiSend/MultiSendContract'
+import { GnosisSafeContract, MultiSendContract, SafeVersion } from '@gnosis.pm/safe-core-sdk-types'
+import { SAFE_LAST_VERSION } from '../contracts/config'
+import {
+  getMultiSendContractDeployment,
+  getSafeContractDeployment
+} from '../contracts/safeDeploymentContracts'
 import { SafeConfig } from '../Safe'
 import { ContractNetworksConfig } from '../types'
 
@@ -30,23 +33,30 @@ class ContractManager {
     const chainId = await ethAdapter.getChainId()
     const customContracts = contractNetworks?.[chainId]
     this.#contractNetworks = contractNetworks
+    this.#isL1SafeMasterCopy = isL1SafeMasterCopy
+
+    let safeSingletonDeployment = getSafeContractDeployment(
+      SAFE_LAST_VERSION,
+      chainId,
+      isL1SafeMasterCopy
+    )
     const temporarySafeContract = ethAdapter.getSafeContract({
       safeVersion: SAFE_LAST_VERSION,
       chainId,
-      isL1SafeMasterCopy,
+      singletonDeployment: safeSingletonDeployment,
       customContractAddress: safeAddress,
       customContractAbi: customContracts?.safeMasterCopyAbi
     })
     if ((await ethAdapter.getContractCode(temporarySafeContract.getAddress())) === '0x') {
       throw new Error('Safe Proxy contract is not deployed in the current network')
     }
-    const safeVersion = (await temporarySafeContract.getVersion()) as SafeVersion
 
-    this.#isL1SafeMasterCopy = isL1SafeMasterCopy
+    const safeVersion = (await temporarySafeContract.getVersion()) as SafeVersion
+    safeSingletonDeployment = getSafeContractDeployment(safeVersion, chainId, isL1SafeMasterCopy)
     const safeContract = ethAdapter.getSafeContract({
       safeVersion,
       chainId,
-      isL1SafeMasterCopy,
+      singletonDeployment: safeSingletonDeployment,
       customContractAddress: safeAddress,
       customContractAbi: customContracts?.safeMasterCopyAbi
     })
@@ -55,9 +65,11 @@ class ContractManager {
     }
     this.#safeContract = safeContract
 
+    const multiSendDeployment = getMultiSendContractDeployment(safeVersion, chainId)
     const multiSendContract = await ethAdapter.getMultiSendContract({
       safeVersion,
       chainId,
+      singletonDeployment: multiSendDeployment,
       customContractAddress: customContracts?.multiSendAddress,
       customContractAbi: customContracts?.multiSendAbi
     })
