+++ packages/safe-core-sdk/src/types/index.ts
@@ -1,29 +1,4 @@
-interface AbiInput {
-  name: string
-  type: string
-  indexed?: boolean
-  components?: AbiInput[]
-  internalType?: string
-}
-
-interface AbiOutput {
-  name: string
-  type: string
-  components?: AbiOutput[]
-  internalType?: string
-}
-
-export interface AbiItem {
-  anonymous?: boolean
-  constant?: boolean
-  inputs?: AbiInput[]
-  name?: string
-  outputs?: AbiOutput[]
-  payable?: boolean
-  stateMutability: string
-  type: string
-  gas?: number
-}
+import { AbiItem } from '@gnosis.pm/safe-core-sdk-types'
 
 export interface ContractNetworkConfig {
   /** multiSendAddress - Address of the MultiSend contract deployed on a specific network */
