+++ packages/safe-core-sdk/src/utils/transactions/gas.ts
@@ -1,7 +1,10 @@
 import { BigNumber } from '@ethersproject/bignumber'
-import { OperationType, SafeTransaction } from '@gnosis.pm/safe-core-sdk-types'
-import EthAdapter from 'ethereumLibs/EthAdapter'
-import GnosisSafeContract from '../../contracts/GnosisSafe/GnosisSafeContract'
+import {
+  EthAdapter,
+  GnosisSafeContract,
+  OperationType,
+  SafeTransaction
+} from '@gnosis.pm/safe-core-sdk-types'
 
 function estimateDataGasCosts(data: string): number {
   const reducer = (accumulator: number, currentValue: string) => {
