+++ packages/safe-core-sdk/src/utils/transactions/types.ts
@@ -1,36 +1,6 @@
-import { ContractTransaction } from '@ethersproject/contracts'
 import { SafeTransactionDataPartial } from '@gnosis.pm/safe-core-sdk-types'
-import { PromiEvent, TransactionReceipt } from 'web3-core/types'
 
 export type SafeTransactionOptionalProps = Pick<
   SafeTransactionDataPartial,
   'safeTxGas' | 'baseGas' | 'gasPrice' | 'gasToken' | 'refundReceiver' | 'nonce'
 >
-
-export interface TransactionOptions {
-  from?: string
-  gas?: number | string
-  gasLimit?: number | string
-  safeTxGas?: number | string
-  gasPrice?: number | string
-}
-
-export interface BaseTransactionResult {
-  hash: string
-}
-
-export interface Web3TransactionResult extends BaseTransactionResult {
-  promiEvent: PromiEvent<TransactionReceipt>
-  options?: TransactionOptions
-}
-
-export interface EthersTransactionResult extends BaseTransactionResult {
-  transactionResponse: ContractTransaction
-  options?: TransactionOptions
-}
-
-export interface TransactionResult extends BaseTransactionResult {
-  promiEvent?: PromiEvent<TransactionReceipt>
-  transactionResponse?: ContractTransaction
-  options?: TransactionOptions
-}
