+++ packages/safe-core-sdk/tsconfig.json
@@ -5,5 +5,5 @@
     "outDir": "dist",
     "baseUrl": "src"
   },
-  "include": ["src", "src/**/*.json", "typechain/src"]
+  "include": ["src", "src/**/*.json"]
 }
